src/requirements.txt
flask
mangum
asgiref
discord-interactions
openai
discord
python-decouple
pytest